Add domain_child module
Add the new module microsoft.ad.domain_child which can be used to create
a child or tree domain in an existing forest.
