+++ plugins/action/domain_child.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2022 Ansible Project
+# Copyright (c) 2024 Ansible Project
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from ..plugin_utils._module_with_reboot import DomainPromotionWithReboot
