+++ plugins/plugin_utils/_module_with_reboot.py
@@ -156,7 +156,7 @@ class ActionModuleWithReboot(ActionBase):
 
                 if self._ad_should_rerun(module_res) and not self._task.check_mode:
                     display.vv(
-                        "Module result has indicated it should rerun after a reboot has occured, rerunning"
+                        "Module result has indicated it should rerun after a reboot has occurred, rerunning"
                     )
                     continue
 
@@ -169,3 +169,38 @@ class ActionModuleWithReboot(ActionBase):
         result = merge_hash(result, module_res)
 
         return self._ad_process_result(result)
+
+
+class DomainPromotionWithReboot(ActionModuleWithReboot):
+    """Domain Promotion Action Plugin with Auto Reboot.
+
+    An action plugin that runs a task that can promote the target Windows host
+    to a domain controller. It implements the common reboot handling for that
+    particular task.
+    """
+
+    def __init__(self, *args: t.Any, **kwargs: t.Any) -> None:
+        super().__init__(*args, **kwargs)
+        self._ran_once = False
+
+    def _ad_should_rerun(self, result: t.Dict[str, t.Any]) -> bool:
+        ran_once = self._ran_once
+        self._ran_once = True
+
+        if ran_once or not result.get("_do_action_reboot", False):
+            return False
+
+        if self._task.check_mode:
+            # Assume that on a rerun it will not have failed and that it
+            # ran successful.
+            result["failed"] = False
+            result.pop("msg", None)
+            return False
+
+        else:
+            return True
+
+    def _ad_process_result(self, result: t.Dict[str, t.Any]) -> t.Dict[str, t.Any]:
+        result.pop("_do_action_reboot", None)
+
+        return result
