+++ tests/integration/targets/domain_child/ansible.cfg
@@ -1,0 +1,4 @@
+[defaults]
+callback_result_format = yaml
+inventory = inventory.yml
+retry_files_enabled = False
